[proliferate] clean anonymous telemetry ownership
