+++ server/tests/unit/test_anonymous_telemetry_store.py
@@ -2,55 +2,41 @@ from __future__ import annotations
 
 import pytest
 from sqlalchemy import select
-from sqlalchemy.ext.asyncio import async_sessionmaker
+from sqlalchemy.ext.asyncio import AsyncSession
 
-from proliferate.db import engine as engine_module
 from proliferate.db.models.anonymous_telemetry import AnonymousTelemetryLocalInstall
 from proliferate.db.store.anonymous_telemetry import load_or_create_local_install_id
 
 
 @pytest.mark.asyncio
 async def test_load_or_create_local_install_id_reuses_existing_surface_id(
-    test_engine,  # type: ignore[no-untyped-def]
+    db_session: AsyncSession,
 ) -> None:
-    original_factory = engine_module.async_session_factory
-    engine_module.async_session_factory = async_sessionmaker(test_engine, expire_on_commit=False)
+    first = await load_or_create_local_install_id(db_session, "server")
+    second = await load_or_create_local_install_id(db_session, "server")
 
-    try:
-        first = await load_or_create_local_install_id("server")
-        second = await load_or_create_local_install_id("server")
+    assert first == second
 
-        assert first == second
-
-        async with engine_module.async_session_factory() as session:
-            records = (
-                (
-                    await session.execute(
-                        select(AnonymousTelemetryLocalInstall).where(
-                            AnonymousTelemetryLocalInstall.surface == "server"
-                        )
-                    )
+    records = (
+        (
+            await db_session.execute(
+                select(AnonymousTelemetryLocalInstall).where(
+                    AnonymousTelemetryLocalInstall.surface == "server"
                 )
-                .scalars()
-                .all()
             )
-        assert len(records) == 1
-        assert records[0].install_uuid == first
-    finally:
-        engine_module.async_session_factory = original_factory
+        )
+        .scalars()
+        .all()
+    )
+    assert len(records) == 1
+    assert records[0].install_uuid == first
 
 
 @pytest.mark.asyncio
 async def test_load_or_create_local_install_id_keeps_surface_ids_distinct(
-    test_engine,  # type: ignore[no-untyped-def]
+    db_session: AsyncSession,
 ) -> None:
-    original_factory = engine_module.async_session_factory
-    engine_module.async_session_factory = async_sessionmaker(test_engine, expire_on_commit=False)
+    desktop_install = await load_or_create_local_install_id(db_session, "desktop")
+    server_install = await load_or_create_local_install_id(db_session, "server")
 
-    try:
-        desktop_install = await load_or_create_local_install_id("desktop")
-        server_install = await load_or_create_local_install_id("server")
-
-        assert desktop_install != server_install
-    finally:
-        engine_module.async_session_factory = original_factory
+    assert desktop_install != server_install
